requirements.txt
absl-py==2.1.0
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.4
action-tutorials-py==0.20.4
actionlib-msgs==4.2.4
ament-cmake-test==1.3.9
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
astunparse==1.6.3
bond==3.0.2
builtin-interfaces==1.2.1
cartographer-ros-msgs==2.0.9002
certifi==2024.6.2
charset-normalizer==3.3.2
composition-interfaces==1.2.1
control-msgs==4.5.0
controller==2023.1.2
controller-manager==2.41.0
controller-manager-msgs==2.41.0
cv-bridge==3.2.1
demo-nodes-py==0.20.4
diagnostic-msgs==4.2.4
diagnostic-updater==3.1.2
domain-coordinator==0.10.0
dwb-msgs==1.1.15
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
flatbuffers==24.3.25
gast==0.5.4
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
image-geometry==3.2.1
interactive-markers==2.3.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
keras==3.3.3
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
libclang==18.1.1
lifecycle-msgs==1.2.1
logging-demo==0.20.4
map-msgs==2.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
message-filters==4.3.4
ml-dtypes==0.3.2
namex==0.0.8
nav-2d-msgs==1.1.15
nav-msgs==4.2.4
nav2-common==1.1.15
nav2-msgs==1.1.15
nav2-simple-commander==1.0.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
osrf-pycommon==2.0.2
packaging==24.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.4
pillow==10.3.0
protobuf==4.25.3
Pygments==2.18.0
python-qt-binding==1.1.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.4
rcl-interfaces==1.2.1
rclpy==3.3.13
rcutils==5.1.6
requests==2.32.3
resource-retriever==3.1.2
rich==13.7.1
rmw-dds-common==1.6.0
ros2action==0.18.10
ros2bag==0.15.11
ros2cli==0.18.10
ros2component==0.18.10
ros2controlcli==2.41.0
ros2doctor==0.18.10
ros2interface==0.18.10
ros2launch==0.19.7
ros2lifecycle==0.18.10
ros2multicast==0.18.10
ros2node==0.18.10
ros2param==0.18.10
ros2pkg==0.18.10
ros2run==0.18.10
ros2service==0.18.10
ros2topic==0.18.10
rosbag2-interfaces==0.15.11
rosbag2-py==0.15.11
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scripts==3.7.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
slam-toolbox==2.6.8
smclib==3.0.2
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
teleop-twist-keyboard==2.4.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tf-transformations==1.0.0
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.7
topic-monitor==0.20.4
trajectory-msgs==4.2.4
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.4.2
typing_extensions==4.12.2
unique-identifier-msgs==2.2.1
urllib3==2.2.2
vehicle==2023.1.2
vision-msgs==4.1.1
visualization-msgs==4.2.4
webots-ros2==2023.1.2
webots-ros2-driver==2023.1.2
webots-ros2-epuck==2023.1.2
webots-ros2-importer==2023.1.2
webots-ros2-mavic==2023.1.2
webots-ros2-msgs==2023.1.2
webots-ros2-tesla==2023.1.2
webots-ros2-tiago==2023.1.2
webots-ros2-turtlebot==2023.1.2
webots-ros2-universal-robot==2023.1.2
werkzeug==3.0.3
wrapt==1.16.0
xacro==2.0.8
